pyproject.toml
[tool.poetry]
name = "ytrends"
version = "0.1.0"
description = ""
authors = ["Aman Rawat <ar.rawataman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
dj-database-url = "^2.1.0"
psycopg = "^3.1.18"
django-autoslug = "^1.9.9"
django-distill = "^3.1.3"
aiohttp = "^3.9.3"
requests-cache = "^1.2.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pandas = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
